pyproject.toml
[project]
name = "github-monitor"
version = "0.1.0"
description = "Monitor GitHub issues, PRs, and comments through polling with GitHub GraphQL API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"nats-py>=2.12.0,<3",
	"cyclopts>=4.2.1,<5",
	"python-dotenv>=1.2.1,<2",
	"structlog>=25.5.0,<26",
	"requests>=2.32.5,<3",
	"pyyaml>=6.0.2,<7",
]

[project.scripts]
github-monitor = "github_monitor.cli.cli:app"

[dependency-groups]
dev = [
    "ruff>=0.14.3,<0.15",
    "pytest>=8.4.2,<9",
    "pytest-asyncio>=1.2,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes (includes unused imports)
    "I",     # isort (import sorting)
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
]

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true
split-on-trailing-comma = false
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-after-imports = 2

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
